Two Hundred Twenty Five Commit: Refactor database schema

+++ store/models.py
@@ -1,10 +1,7 @@
-from django.contrib.auth.models import User
+from django.conf import settings
 from django.db import models
 from django.urls import reverse
-from django.conf import settings
 
-
-# Create your models here.
 class ProductManager(models.Manager):
     def get_queryset(self):
         return super(ProductManager, self).get_queryset().filter(is_active=True)
@@ -32,7 +29,7 @@ class Product(models.Model):
     description = models.TextField(blank=True)
     image = models.ImageField(upload_to='images/', default='images/default.png')
     slug = models.SlugField(max_length=255)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=4, decimal_places=2)
     in_stock = models.BooleanField(default=True)
     is_active = models.BooleanField(default=True)
     created = models.DateTimeField(auto_now_add=True)
@@ -49,3 +46,5 @@ class Product(models.Model):
 
     def __str__(self):
         return self.title
+
+
